Branch for getting structs right for ABF2 files
Need to get the relevant common struct definitions from ABF1 headers (like ABF
tags and data sections).

Thinking that the c library should deserialize the ABF format as stream of
bytes to deal explicitly with endian-ness and packing. The structs as given in
the ABF2 documentation use the #pragma pack(1) directive to give non-aligned
structure members. This pragma is not standard, although it is part of gcc and
Microsoft Visual C++ (the two main target compilers). Thinking that the
library should stay as platform-independent as possible, though.